+++ engine/apps/public_api/serializers/escalation_policies.py
@@ -7,7 +7,7 @@ from rest_framework import fields, serializers
 from apps.alerts.models import EscalationChain, EscalationPolicy
 from apps.schedules.models import OnCallSchedule
 from apps.slack.models import SlackUserGroup
-from apps.user_management.models import User
+from apps.user_management.models import Team, User
 from apps.webhooks.models import Webhook
 from common.api_helpers.custom_fields import (
     OrganizationFilteredPrimaryKeyRelatedField,
@@ -57,6 +57,11 @@ class EscalationPolicySerializer(EagerLoadingMixin, OrderedModelSerializer):
         required=False,
         source="notify_to_users_queue",
     )
+    team_to_notify = OrganizationFilteredPrimaryKeyRelatedField(
+        queryset=Team.objects,
+        required=False,
+        source="notify_to_team_members",
+    )
     persons_to_notify_next_each_time = UsersFilteredByOrganizationField(
         queryset=User.objects,
         required=False,
@@ -95,6 +100,7 @@ class EscalationPolicySerializer(EagerLoadingMixin, OrderedModelSerializer):
             "duration",
             "important",
             "persons_to_notify",
+            "team_to_notify",
             "persons_to_notify_next_each_time",
             "notify_on_call_from_schedule",
             "group_to_notify",
@@ -154,6 +160,7 @@ class EscalationPolicySerializer(EagerLoadingMixin, OrderedModelSerializer):
     def _get_field_to_represent(self, step, result):
         fields_to_remove = [
             "duration",
+            "team_to_notify",
             "persons_to_notify",
             "persons_to_notify_next_each_time",
             "notify_on_call_from_schedule",
@@ -174,6 +181,11 @@ class EscalationPolicySerializer(EagerLoadingMixin, OrderedModelSerializer):
             EscalationPolicy.STEP_NOTIFY_MULTIPLE_USERS_IMPORTANT,
         ]:
             fields_to_remove.remove("persons_to_notify")
+        elif step in [
+            EscalationPolicy.STEP_NOTIFY_TEAM_MEMBERS,
+            EscalationPolicy.STEP_NOTIFY_TEAM_MEMBERS_IMPORTANT,
+        ]:
+            fields_to_remove.remove("team_to_notify")
         elif step == EscalationPolicy.STEP_NOTIFY_USERS_QUEUE:
             fields_to_remove.remove("persons_to_notify_next_each_time")
         elif step in [EscalationPolicy.STEP_NOTIFY_GROUP, EscalationPolicy.STEP_NOTIFY_GROUP_IMPORTANT]:
@@ -203,6 +215,7 @@ class EscalationPolicySerializer(EagerLoadingMixin, OrderedModelSerializer):
             "wait_delay",
             "notify_schedule",
             "notify_to_group",
+            "notify_to_team_members",
             "custom_button_trigger",
             "custom_webhook",
             "from_time",
@@ -229,6 +242,8 @@ class EscalationPolicySerializer(EagerLoadingMixin, OrderedModelSerializer):
             validated_data_fields_to_remove.remove("notify_to_users_queue")
         elif step in [EscalationPolicy.STEP_NOTIFY_GROUP, EscalationPolicy.STEP_NOTIFY_GROUP_IMPORTANT]:
             validated_data_fields_to_remove.remove("notify_to_group")
+        elif step in [EscalationPolicy.STEP_NOTIFY_TEAM_MEMBERS, EscalationPolicy.STEP_NOTIFY_TEAM_MEMBERS_IMPORTANT]:
+            validated_data_fields_to_remove.remove("notify_to_team_members")
         elif step == EscalationPolicy.STEP_TRIGGER_CUSTOM_BUTTON:
             validated_data_fields_to_remove.remove("custom_button_trigger")
         elif step == EscalationPolicy.STEP_TRIGGER_CUSTOM_WEBHOOK:
@@ -280,6 +295,11 @@ class EscalationPolicyUpdateSerializer(EscalationPolicySerializer):
                     EscalationPolicy.STEP_NOTIFY_MULTIPLE_USERS_IMPORTANT,
                 ]:
                     instance.notify_to_users_queue.clear()
+                if step not in [
+                    EscalationPolicy.STEP_NOTIFY_TEAM_MEMBERS,
+                    EscalationPolicy.STEP_NOTIFY_TEAM_MEMBERS_IMPORTANT,
+                ]:
+                    instance.notify_to_team_members = None
                 if step not in [EscalationPolicy.STEP_NOTIFY_GROUP, EscalationPolicy.STEP_NOTIFY_GROUP_IMPORTANT]:
                     instance.notify_to_group = None
                 if step != EscalationPolicy.STEP_TRIGGER_CUSTOM_BUTTON: