+++ grafana-plugin/src/components/Policy/EscalationPolicy.tsx
@@ -20,6 +20,7 @@ import {
   EscalationPolicy as EscalationPolicyType,
   EscalationPolicyOption,
 } from 'models/escalation_policy/escalation_policy.types';
+import { GrafanaTeam } from 'models/grafana_team/grafana_team.types';
 import { OutgoingWebhook } from 'models/outgoing_webhook/outgoing_webhook.types';
 import { Schedule } from 'models/schedule/schedule.types';
 import { User } from 'models/user/user.types';
@@ -109,6 +110,8 @@ class _EscalationPolicy extends React.Component<EscalationPolicyProps, any> {
         return this.renderWaitDelays();
       case 'slack_user_group':
         return this.renderNotifyUserGroup();
+      case 'team':
+        return this.renderNotifyTeam();
       case 'schedule':
         return this.renderNotifySchedule();
       case 'custom_webhook':
@@ -415,6 +418,33 @@ class _EscalationPolicy extends React.Component<EscalationPolicyProps, any> {
             const webhook = outgoingWebhookStore.items[id];
             return webhook.trigger_type_name === 'Escalation step';
           }}
+        />
+      </WithPermissionControlTooltip>
+    );
+  }
+
+  renderNotifyTeam() {
+    const {
+      data,
+      isDisabled,
+      store: { grafanaTeamStore },
+    } = this.props;
+    const { notify_to_team_members } = data;
+
+    return (
+      <WithPermissionControlTooltip key="notify_to_team_members" userAction={UserActions.EscalationChainsWrite}>
+        <GSelect<GrafanaTeam>
+          disabled={isDisabled}
+          items={grafanaTeamStore.items}
+          fetchItemsFn={grafanaTeamStore.updateItems}
+          getSearchResult={grafanaTeamStore.getSearchResult}
+          displayField="name"
+          valueField="id"
+          placeholder="Select Team"
+          className={cx('select', 'control')}
+          value={notify_to_team_members}
+          onChange={this.getOnChangeHandler('notify_to_team_members')}
+          width={'auto'}
         />
       </WithPermissionControlTooltip>
     );